requirements-minimal.txt
numpy>=1.23.2,<2.0
pandas==2.2.3
requests==2.32.3
websockets==12.0
pydantic==2.8.2
pydantic-settings==2.1.0
python-dotenv==1.0.1
streamlit==1.37.1
plotly==5.23.0
tqdm==4.66.5
google-cloud-bigquery==3.15.0
google-cloud-functions==1.20.4
google-cloud-scheduler==2.15.0
pyarrow==21.0.0
ray[default]==2.49.2
aiohttp==3.9.1
snscrape==0.6.2.20230320
vaderSentiment==3.3.2
feedparser==6.0.10
TA-Lib==0.4.28
scikit-learn==1.3.2
xgboost==2.0.3
lightgbm==4.1.0
cryptography==41.0.7
matplotlib==3.8.2
bandit
safety
pre-commit
